Move extensions into another file
